requirements.txt
earthengine-api
pandas
geopy
tqdm
openpyxl
python-dotenv
pytest
pytest-cov